Moved physics equations and radiation to physics folder
