+++ src/jaff/physics/__init__.py
@@ -1,0 +1,3 @@
+from . import constants, equations, radiation
+
+__all__ = ["constants", "equations", "radiation"]
